+++ tracer/src/instruction/remw.rs
@@ -15,15 +15,18 @@ declare_riscv_instr!(
 
 impl REMW {
     fn exec(&self, cpu: &mut Cpu, _: &mut <REMW as RISCVInstruction>::RAMAccess) {
+        // REMW and REMUW are RV64 instructions that provide the corresponding signed and unsigned
+        // remainder operations. Both REMW and REMUW always sign-extend the 32-bit result to 64 bits,
+        // including on a divide by zero.
         let dividend = cpu.x[self.operands.rs1] as i32;
         let divisor = cpu.x[self.operands.rs2] as i32;
-        cpu.x[self.operands.rd] = if divisor == 0 {
-            dividend as i64
+        cpu.x[self.operands.rd] = (if divisor == 0 {
+            dividend
         } else if dividend == i32::MIN && divisor == -1 {
             0
         } else {
-            (dividend.wrapping_rem(divisor)) as i64
-        };
+            dividend.wrapping_rem(divisor)
+        }) as i64;
     }
 }
 impl RISCVTrace for REMW {}
